Delete temporary files for chunk streams (on close) + Memory channel buffer (with threshold)
- Add the configuration property (play.netty.memoryThreshold)
- Super class InputStreamChannelBuffer
- Careful : need to close the body on request to delete the associated
temporary file
