+++ framework/src/play/server/HttpServerPipelineFactory.java
@@ -1,26 +1,27 @@
     package play.server;
 
+import static org.jboss.netty.channel.Channels.pipeline;
+
 import org.jboss.netty.channel.ChannelPipeline;
 import org.jboss.netty.channel.ChannelPipelineFactory;
-import org.jboss.netty.handler.codec.http.HttpChunkAggregator;
 import org.jboss.netty.handler.codec.http.HttpRequestDecoder;
 import org.jboss.netty.handler.codec.http.HttpResponseEncoder;
-import play.Play;
 
-import static org.jboss.netty.channel.Channels.pipeline;
+import play.Play;
 
 public class HttpServerPipelineFactory implements ChannelPipelineFactory {
 
     public ChannelPipeline getPipeline() throws Exception {
 
-        Integer max = Integer.valueOf(Play.configuration.getProperty("play.netty.maxContentLength", "-1"));
+        Integer max 		= Integer.valueOf(Play.configuration.getProperty("play.netty.maxContentLength", "-1"));
+        Integer threshold 	= Integer.valueOf(Play.configuration.getProperty("play.netty.memoryThreshold", 	"1024")); //1kB by default
            
         ChannelPipeline pipeline = pipeline();
         PlayHandler playHandler = new PlayHandler();
         
         pipeline.addLast("flashPolicy", new FlashPolicyHandler()); 
         pipeline.addLast("decoder", new HttpRequestDecoder());
-        pipeline.addLast("aggregator", new StreamChunkAggregator(max));
+        pipeline.addLast("aggregator", new StreamChunkAggregator(max, threshold));
         pipeline.addLast("encoder", new HttpResponseEncoder());
         pipeline.addLast("chunkedWriter", playHandler.chunkedWriteHandler);
         pipeline.addLast("handler", playHandler);
