+++ framework/src/play/server/ssl/SslHttpServerPipelineFactory.java
@@ -20,8 +20,9 @@ public class SslHttpServerPipelineFactory implements ChannelPipelineFactory {
 
     public ChannelPipeline getPipeline() throws Exception {
 
-        Integer max = Integer.valueOf(Play.configuration.getProperty("play.netty.maxContentLength", "-1"));
-        String mode = Play.configuration.getProperty("play.netty.clientAuth", "none");
+        Integer max 		= Integer.valueOf(Play.configuration.getProperty("play.netty.maxContentLength", "-1"));
+        Integer threshold 	= Integer.valueOf(Play.configuration.getProperty("play.netty.memoryThreshold", 	"1024"));  //1kB by default
+        String mode 		= Play.configuration.getProperty("play.netty.clientAuth", "none");
 
         ChannelPipeline pipeline = pipeline();
 
@@ -40,7 +41,7 @@ public class SslHttpServerPipelineFactory implements ChannelPipelineFactory {
         pipeline.addLast("flashPolicy", new FlashPolicyHandler());
         pipeline.addLast("ssl", new SslHandler(engine));
         pipeline.addLast("decoder", new HttpRequestDecoder());
-        pipeline.addLast("aggregator", new StreamChunkAggregator(max));
+        pipeline.addLast("aggregator", new StreamChunkAggregator(max, threshold));
         pipeline.addLast("encoder", new HttpResponseEncoder());
         pipeline.addLast("chunkedWriter", new ChunkedWriteHandler());
